combined_spaces_interface/requirements.txt
transformers
torch
newspaper3k
gradio